# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/treatments": {
      "post": {
        "operationId": "TreatmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos requeridos para crear un tratamiento",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTreatmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tratamiento creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          }
        },
        "summary": "Crear un nuevo tratamiento",
        "tags": [
          "Treatments"
        ]
      },
      "get": {
        "operationId": "TreatmentsController_findPaginated",
        "parameters": [
          {
            "name": "term",
            "required": false,
            "in": "query",
            "description": "Término de búsqueda opcional.",
            "schema": {
              "example": "ejemplo",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Número de elementos por página.",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Dirección de ordenamiento (ASC/DESC)",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Texto para búsqueda",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Límite de resultados por página",
            "schema": {
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tratamientos paginados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationDto"
                }
              }
            }
          }
        },
        "summary": "Obtener tratamientos paginados",
        "tags": [
          "Treatments"
        ]
      }
    },
    "/treatments/calculate": {
      "post": {
        "description": "Calcula los requerimientos técnicos para un tratamiento térmico basado en parámetros de flujo y composición",
        "operationId": "TreatmentsController_calculate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos de entrada para cálculo de tratamiento",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateTreatmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultados del cálculo técnico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentCalculationsDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos o fuera de rango"
          }
        },
        "summary": "Calcular parámetros de tratamiento",
        "tags": [
          "Treatments"
        ]
      }
    },
...
    "/treatments/{id}": {
      "get": {
        "operationId": "TreatmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tratamiento",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tratamiento encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tratamiento no encontrado"
          }
        },
        "summary": "Obtener un tratamiento por ID",
        "tags": [
          "Treatments"
        ]
      },
      "patch": {
        "operationId": "TreatmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tratamiento a actualizar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Datos a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTreatmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tratamiento actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "404": {
            "description": "Tratamiento no encontrado"
          }
        },
        "summary": "Actualizar un tratamiento",
        "tags": [
          "Treatments"
        ]
      },
      "delete": {
        "operationId": "TreatmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tratamiento a eliminar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tratamiento eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tratamiento no encontrado"
          }
        },
        "summary": "Eliminar un tratamiento (soft delete)",
        "tags": [
          "Treatments"
        ]
      }
    },
    "/treatments/user/{userId}/all": {
      "get": {
        "operationId": "TreatmentsController_findAllByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID del usuario",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tratamientos del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TreatmentResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "summary": "Obtener todos los tratamientos de un usuario",
        "tags": [
          "Treatments"
        ]
      }
    },
    "/treatments/user/{userId}": {
      "get": {
        "operationId": "TreatmentsController_findPaginatedByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID del usuario",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "term",
            "required": false,
            "in": "query",
            "description": "Término de búsqueda opcional.",
            "schema": {
              "example": "ejemplo",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página para la paginación (base 1).",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Número de elementos por página.",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar los resultados.",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento (ASC o DESC).",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filtrar por tipo de tratamiento",
            "schema": {
              "example": "Deshidratación",
              "type": "string"
            }
          },
          {
            "name": "minTotalFlow",
            "required": false,
            "in": "query",
            "description": "Filtrar por flujo total mínimo (bpd)",
            "schema": {
              "example": 1000,
              "type": "number"
            }
          },
          {
            "name": "maxTotalFlow",
            "required": false,
            "in": "query",
            "description": "Filtrar por flujo total máximo (bpd)",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "minWaterFraction",
            "required": false,
            "in": "query",
            "description": "Filtrar por fracción de agua mínima",
            "schema": {
              "example": 0.1,
              "type": "number"
            }
          },
          {
            "name": "maxWaterFraction",
            "required": false,
            "in": "query",
            "description": "Filtrar por fracción de agua máxima",
            "schema": {
              "example": 0.5,
              "type": "number"
            }
          },
          {
            "name": "minApiGravity",
            "required": false,
            "in": "query",
            "description": "Filtrar por gravedad API mínima",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "maxApiGravity",
            "required": false,
            "in": "query",
            "description": "Filtrar por gravedad API máxima",
            "schema": {
              "example": 40,
              "type": "number"
            }
          },
          {
            "name": "deleted",
            "required": false,
            "in": "query",
            "description": "Filtrar por estado de eliminación lógica",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "createdById",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de usuario creador",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "minCreatedAt",
            "required": false,
            "in": "query",
            "description": "Filtrar por fecha mínima de creación (formato YYYY-MM-DD)",
            "schema": {
              "example": "2023-01-01",
              "type": "string"
            }
          },
          {
            "name": "maxCreatedAt",
            "required": false,
            "in": "query",
            "description": "Filtrar por fecha máxima de creación (formato YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tratamientos paginados del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "summary": "Obtener tratamientos paginados de un usuario",
        "tags": [
          "Treatments"
        ]
      }
    },
...
  },
  "info": {
    "title": "TermoLab API",
    "description": "API para gestión de tratamientos térmicos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "Array de elementos para la página actual.",
            "items": {}
          },
          "paginationData": {
            "description": "Metadatos de paginación.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationDataDto"
              }
            ]
          }
        },
        "required": [
          "data",
          "paginationData"
        ]
      },
...
      "CreateTreatmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del tratamiento",
            "example": "Tratamiento para crudo 18°API"
          },
          "description": {
            "type": "string",
            "description": "Descripción del tratamiento",
            "example": "Diseño para 500 bbl/día con 20% agua"
          },
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "totalFlow": {
            "type": "number",
            "description": "Flujo total de emulsión (bpd)",
            "example": 500
          },
          "waterFraction": {
            "type": "number",
            "description": "Fracción de agua (%)",
            "example": 20
          },
          "inputTemperature": {
            "type": "number",
            "description": "Temperatura de entrada (°F)",
            "example": 75
          },
          "treatmentTemperature": {
            "type": "number",
            "description": "Temperatura de tratamiento (°F)",
            "example": 140
          },
          "ambientTemperature": {
            "type": "number",
            "description": "Temperatura ambiente (°F)",
            "example": 30
          },
          "oilRetentionTime": {
            "type": "number",
            "description": "Tiempo retención petróleo (min)",
            "example": 60
          },
          "waterRetentionTime": {
            "type": "number",
            "description": "Tiempo retención agua (min)",
            "example": 30
          },
          "windSpeed": {
            "type": "number",
            "description": "Velocidad del viento (mph)",
            "example": 15
          },
          "apiGravity": {
            "type": "number",
            "description": "Gravedad API del crudo",
            "example": 18
          }
        },
        "required": [
          "name",
          "type",
          "totalFlow",
          "waterFraction",
          "inputTemperature",
          "treatmentTemperature",
          "ambientTemperature",
          "oilRetentionTime",
          "waterRetentionTime",
          "windSpeed",
          "apiGravity"
        ]
      },
      "TreatmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identificador único de la entidad.",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de creación del registro.",
            "example": "2023-10-27T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de la última actualización del registro.",
            "example": "2023-10-27T10:30:00Z"
          },
          "status": {
            "type": "boolean",
            "description": "Estado del registro (activo/inactivo).",
            "example": true
          },
          "deleted": {
            "type": "boolean",
            "description": "Indicador de borrado lógico.",
            "example": false
          },
          "name": {
            "type": "string",
            "description": "Nombre del tratamiento",
            "example": "Tratamiento para crudo 18°API"
          },
          "description": {
            "type": "string",
            "description": "Descripción del tratamiento",
            "example": "Diseño para 500 bbl/día con 20% agua"
          },
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "totalFlow": {
            "type": "number",
            "description": "Flujo total de emulsión (bpd)",
            "example": 500
          },
          "waterFraction": {
            "type": "number",
            "description": "Fracción de agua (%)",
            "example": 20
          },
          "inputTemperature": {
            "type": "number",
            "description": "Temperatura de entrada (°F)",
            "example": 75
          },
          "treatmentTemperature": {
            "type": "number",
            "description": "Temperatura de tratamiento (°F)",
            "example": 140
          },
          "ambientTemperature": {
            "type": "number",
            "description": "Temperatura ambiente (°F)",
            "example": 30
          },
          "oilRetentionTime": {
            "type": "number",
            "description": "Tiempo retención petróleo (min)",
            "example": 60
          },
          "waterRetentionTime": {
            "type": "number",
            "description": "Tiempo retención agua (min)",
            "example": 30
          },
          "windSpeed": {
            "type": "number",
            "description": "Velocidad del viento (mph)",
            "example": 15
          },
          "apiGravity": {
            "type": "number",
            "description": "Gravedad API del crudo",
            "example": 18
          },
          "calculatedOilFlow": {
            "type": "number",
            "description": "Flujo de petróleo calculado (bpd)",
            "example": 400
          },
          "calculatedWaterFlow": {
            "type": "number",
            "description": "Flujo de agua calculado (bpd)",
            "example": 100
          },
          "oilRetentionVolume": {
            "type": "number",
            "description": "Volumen retención petróleo (bbl)",
            "example": 16.67
          },
          "waterRetentionVolume": {
            "type": "number",
            "description": "Volumen retención agua (bbl)",
            "example": 2.08
          },
          "requiredHeat": {
            "type": "number",
            "description": "Calor requerido (BTU/hr)",
            "example": 262210
          },
          "heatLoss": {
            "type": "number",
            "description": "Pérdidas de calor (BTU/hr)",
            "example": 87120
          },
          "totalHeat": {
            "type": "number",
            "description": "Calor total (BTU/hr)",
            "example": 349330
          },
          "selectedDiameter": {
            "type": "number",
            "description": "Diámetro seleccionado (ft)",
            "example": 4
          },
          "selectedLength": {
            "type": "number",
            "description": "Longitud seleccionada (ft)",
            "example": 15
          },
          "designPressure": {
            "type": "number",
            "description": "Presión de diseño (psig)",
            "example": 50
          },
          "createdBy": {
            "description": "Usuario creador del tratamiento",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "type",
          "totalFlow",
          "waterFraction",
          "inputTemperature",
          "treatmentTemperature",
          "ambientTemperature",
          "oilRetentionTime",
          "waterRetentionTime",
          "windSpeed",
          "apiGravity",
          "calculatedOilFlow",
          "calculatedWaterFlow",
          "oilRetentionVolume",
          "waterRetentionVolume",
          "requiredHeat",
          "heatLoss",
          "totalHeat",
          "selectedDiameter",
          "selectedLength",
          "designPressure",
          "createdBy"
        ]
      },
      "CalculateTreatmentDto": {
        "type": "object",
        "properties": {
          "totalFlow": {
            "type": "number",
            "description": "Flujo total de emulsión (bbl/día)",
            "example": 500
          },
          "waterFraction": {
            "type": "number",
            "description": "Fracción de agua (%)",
            "example": 20
          },
          "inletTemperature": {
            "type": "number",
            "description": "Temperatura de entrada (°F)",
            "example": 75
          },
          "targetTemperature": {
            "type": "number",
            "description": "Temperatura de tratamiento (°F)",
            "example": 140
          },
          "ambientTemperature": {
            "type": "number",
            "description": "Temperatura ambiente (°F)",
            "example": 30
          },
          "oilRetentionTime": {
            "type": "number",
            "description": "Tiempo de retención del petróleo (minutos)",
            "example": 60
          },
          "waterRetentionTime": {
            "type": "number",
            "description": "Tiempo de retención del agua (minutos)",
            "example": 30
          },
          "windSpeed": {
            "type": "number",
            "description": "Velocidad del viento (mph)",
            "example": 15
          },
          "apiGravity": {
            "type": "number",
            "description": "Gravedad API del crudo",
            "example": 18
          }
        },
        "required": [
          "totalFlow",
          "waterFraction",
          "inletTemperature",
          "targetTemperature",
          "ambientTemperature",
          "oilRetentionTime",
          "waterRetentionTime",
          "windSpeed",
          "apiGravity"
        ]
      },
...
      "UpdateTreatmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del tratamiento",
            "example": "Tratamiento para crudo 18°API"
          },
          "description": {
            "type": "string",
            "description": "Descripción del tratamiento",
            "example": "Diseño para 500 bbl/día con 20% agua"
          },
          "type": {
            "type": "string",
            "description": "Tipo de tratador (vertical/horizontal)",
            "example": "horizontal",
            "enum": [
              "vertical",
              "horizontal"
            ]
          },
          "totalFlow": {
            "type": "number",
            "description": "Flujo total de emulsión (bpd)",
            "example": 500
          },
          "waterFraction": {
            "type": "number",
            "description": "Fracción de agua (%)",
            "example": 20
          },
          "inputTemperature": {
            "type": "number",
            "description": "Temperatura de entrada (°F)",
            "example": 75
          },
          "treatmentTemperature": {
            "type": "number",
            "description": "Temperatura de tratamiento (°F)",
            "example": 140
          },
          "ambientTemperature": {
            "type": "number",
            "description": "Temperatura ambiente (°F)",
            "example": 30
          },
          "oilRetentionTime": {
            "type": "number",
            "description": "Tiempo retención petróleo (min)",
            "example": 60
          },
          "waterRetentionTime": {
            "type": "number",
            "description": "Tiempo retención agua (min)",
            "example": 30
          },
          "windSpeed": {
            "type": "number",
            "description": "Velocidad del viento (mph)",
            "example": 15
          },
          "apiGravity": {
            "type": "number",
            "description": "Gravedad API del crudo",
            "example": 18
          }
        }
      },
...
    }
  }
}